requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
Pillow==10.2.0
twilio==8.10.0
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
